Extract useWebsocketHeartbeat hook
